# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bauk Cheescake",
        "version": "0.0.11",
        "description": "appiid: 3232f21e-653e-4e98-b46d-c34cbc751182 - TODO: All users can post new addresses"
    },
    "servers": [
        {
            "url": "https://localhost:443"
        }
    ],
    "security": [
        {
            "BearerAuth": [],
            "Refresh": [],
            "MailAuth": [],
            "AppiId": []
        }
    ],
    "paths": {
...
        "/visits/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Get visit by id",
                "tags": [
                    "Visits"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VisitOrder"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "description": "Put visit",
                "tags": [
                    "Visits"
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostVisit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/visits/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "description": "Get all visits",
                "tags": [
                    "Visits"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "Application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VisitOrder"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "description": "Post visit",
                "tags": [
                    "Visits"
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostVisit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
        "/visits/multi": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "AppiId": []
                    }
                ],
                "description": "Post visits",
                "tags": [
                    "Visits"
                ],
                "requestBody": {
                    "content": {
                        "Application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/PostVisit"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "Refresh": {
                "type": "apiKey",
                "in": "header",
                "name": "Refresh"
            },
            "MailAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "MailToken"
            },
            "AppiId": {
                "type": "apiKey",
                "in": "header",
                "name": "appiId"
            }
        },
        "schemas": {
            "VisitOrder": {
                "required": [
                    "Visit",
                    "Orders"
                ],
                "properties": {
                    "Visit": {
                        "type": "object",
                        "$ref": "#/components/schemas/Visit"
                    },
                    "Orders": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "$ref": "#/components/schemas/Order"
                        }
                    }
                }
            },
            "Visit": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "integer"
                    },
                    "Date": {
                        "type": "string",
                        "format": "date"
                    },
                    "SlotsNumber": {
                        "type": "integer"
                    },
                    "Price": {
                        "type": "integer"
                    },
                    "PlaceName": {
                        "type": "string"
                    },
                    "AreaCode": {
                        "type": "string"
                    },
                    "Active": {
                        "type": "boolean"
                    }
                }
            },
            "PostVisit": {
                "type": "object",
                "properties": {
                    "Date": {
                        "type": "string",
                        "format": "date"
                    },
                    "IdPlace": {
                        "type": "integer"
                    },
                    "SlotsNumber": {
                        "type": "integer"
                    }
                }
            },
            "Order": {
                "type": "object",
                "properties": {
                    "ID": {
                        "type": "integer"
                    },
                    "Price": {
                        "type": "integer"
                    },
                    "Description": {
                        "type": "string"
                    },
                    "Name": {
                        "type": "string"
                    },
                    "FirstName": {
                        "type": "string"
                    },
                    "LastName": {
                        "type": "string"
                    },
                    "PhoneNumber": {
                        "type": "string"
                    },
                    "Email": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}